split searchdb.py
