+++ backend/util/__init__.py
@@ -1,3 +1,2 @@
-from .searchdb import *
 from .cookie import *
 from .validator import *
